+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Expose internal errors [#23](https://github.com/rokwire/logging-library-go/issues/23)
+
 ## [2.1.0] - 2022-12-01
 ### Added
 - Regex in HTTPRequestProperties [#20](https://github.com/rokwire/logging-library-go/issues/20)